Fix 5 build errors and 8 warnings
Errors:
- CS0128: Rename duplicate 'noMaterial' variable to 'noMaterial2' in StructuralDesignSuite.cs
- CS1061: Replace GetWorksetCount() with GetWorksets().Count in ModelEngine.cs
- CS1503: Fix FormulaEngine.LoadFormulas() — pass CSV path string instead of Document
- CS1061: Replace SpatialElementTag.Room with RoomTag.Room in WarningsManager.cs
- CS0618: Use ElementId(long) instead of ElementId(int) in Phase75Enhancements.cs

Warnings:
- CS8632: Remove nullable annotation (XYZ?) in StructuralIntelligenceEngine.cs
- CS0219: Remove unused stiffSumX/stiffSumY in StructuralPrecisionEngine.cs
- CS0219: Remove unused hsc variable in StructuralDesignSuite.cs
- CS0219: Remove unused isCantilever in StructuralDeepEngine.cs
- CS0219: Remove unused roomCorrection in MEPIntelligenceEngine.cs

+++ StingTools/Core/Phase75Enhancements.cs
@@ -805,7 +805,7 @@ namespace StingTools.Core
 
             foreach (var kvp in rootCauseElements)
             {
-                var elementId = new ElementId((int)kvp.Key);
+                var elementId = new ElementId((long)kvp.Key);
                 var relatedWarnings = report.Warnings
                     .Where(w => w.FailingElements != null && w.FailingElements.Contains(elementId))
                     .ToList();
